% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vietnam63.R
\docType{package}
\name{vietnam63}
\alias{vietnam63}
\alias{vietnam63-package}
\title{vietnam63: Shape files of the communes, districts and provinces of Vietnam as of after 2008}
\description{
The vietnam63 package provides the polygons of the communes, districts and provinces of Vietnam as defined since 2008.
}

